release_android: rework add_commit_msg to use generic names
add_commit_msg aims to add elements to an hashmap where the key is the
path and the value passed as argument to the function.
This function can be generic and should not contains any reference to
commits, title, mtk_android_out ...

add_commit_msg has been renamed add_to_path_hashmap and all variables
involved have now a generic name.

The arguments order has also been changed, the path is passed as 2nd
argument and the value as 3th.

Signed-off-by: Julien Masson <jmasson@baylibre.com>