release_android: add support of extra_projects yaml entry
The support of the new yaml entry android.extra_projects has been
added into release_android script.
This optional yaml entry allow the user to list projects that will be
added to from_projects list.
Initially from_projects list contains PROJECTS_AIOT and this list can
be appended if the user set android.extra_projects in the
corresponding yaml config.
from_projects list will be used by commit_binaries to show
informations of each project in the commit message of the binaries
project.

A typical use case is when we want to show informations from specific
projects for a customer and these informations must be added only in
the commit message of the customer binaries project.

Signed-off-by: Julien Masson <jmasson@baylibre.com>