[REVERT-ME] remove i500 pumpkin config
i500 pumpkin doesn't build anymore due to latest changes on libatf.
A lot of work needs to be done to continue the support of this board.

Our CI is broken since this config doesn't build and thus all other
bootloaders binaries cannot be generated.
For now let's remove the i500 pumpkin config, the support will be
added later.

Signed-off-by: Julien Masson <jmasson@baylibre.com>