[WORKAROUND] u-boot: spl-da: save space by disable CONFIG_SPL_POWER
With recent changes in the U-Boot SPL, when we build it as Download
Agent we exceed the maximum size for i350 (lowest maximum size
compared to others SoC).

In order to save some space we can disable CONFIG_SPL_POWER.
This config is not needed because ROM code (BL1) has already setup
regulators, pmic ...

However that is a workaround, we may need to have this config enabled
in the future (enable/disable regulators ...).

Signed-off-by: Julien Masson <jmasson@baylibre.com>