release_android: clean only optee-ta output directory
When we run release_android script, we build binaries for release and
userdebug mode.
We should not remove the whole output directory, otherwise the output
directory will only contains the binaries from the last build.
We need to have both binaries: release and userdebug.

Except for the optee-ta binaries, we only use the release binaries.
This folder can be cleaned before copying all the binaries.

Fixes: d8a121a release_android: remove all previous build files
Signed-off-by: Julien Masson <jmasson@baylibre.com>