treewide: rework plat handling
The plat yaml entry was a global var used for several bootloader
components: OP-TEE, BL31, libatf ...
This logic works only if all the projects use the plat name.

With the board i350 EVK we were lucky because all projects are aligned
with a plat set as mt8365 except for OP-TEE (mt8175).

However this logic has limitation, if one of these projects use a
different plat name, we had to add condition to handle specific case.
That can start to be mess if more projects are not aligned with the
global plat var.

This commit add and handle plat yaml entry for each bootloader
components.

In that way the build_*.sh won't need to be updated to handle all the
different cases for future boards.

Each board specify the right plat name for each bootloader components
in their yaml config.

Signed-off-by: Julien Masson <jmasson@baylibre.com>